Add SwapVar module to swap around the Coveo variable, and use it everywhere (#204)
https://coveord.atlassian.net/browse/JSUI-1199
