+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "coveo-search-ui",
-  "version": "0.0.31",
+  "version": "1.0.0",
   "description": "Coveo JavaScript Search Framework",
   "main": "./bin/js/CoveoJsSearch.js",
   "scripts": {
