+++ pages/GoogleDrive.html
@@ -8,7 +8,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, height=device-height, initial-scale=1.0,  maximum-scale=1.0, user-scalable=0">
     <script src="js/CoveoJsSearch.js"></script>
-    <script src="js/CoveoJsSearch.Dependencies.js"></script>
     <script src="js/templates/templatesNew.js"></script>
     <link rel="stylesheet" href="css/CoveoFullSearchNewDesign.css" />
     <script>
