+++ pages/Recommendation.html
@@ -6,7 +6,6 @@
     <title>Recommendation</title> 
     <link rel="stylesheet" href="./css/CoveoFullSearchNewDesign.css"/>
     <script src="js/CoveoJsSearch.js"></script>
-    <script src="js/CoveoJsSearch.Dependencies.js"></script>
     <script src="js/templates/templatesNew.js"></script>
     <script>
       document.addEventListener('DOMContentLoaded', function(){
