+++ pages/Standalone.html
@@ -5,8 +5,7 @@
     <meta name="viewport" content="width=device-width, height=device-height" >
     <title>Title</title> 
     <link rel="stylesheet" href="./css/CoveoFullSearchNewDesign.css"/>
-    <script src="js/CoveoJsSearch.Dependencies.js"></script>
-    <script src="js/CoveoJsSearch.js"></script>
+    <script src="js/CoveoJsSearch.Searchbox.js"></script>
     <script src="js/templates/templatesNew.js"></script>
     <script>
       document.addEventListener('DOMContentLoaded', function(){
