+++ plop/plopTemplates/plop.bundle.template.hbs
@@ -12,17 +12,5 @@ export * from '../src/TemplatesModules';
 export { {{this}} } from '../src/ui/{{this}}/{{this}}.ts';
 {{/each}}
 
-// Webpack output a library target with a temporary name.
-// This is to allow end user to put CoveoJsSearch.Dependencie.js before or after the main CoveoJsSearch.js, without breaking
-// This code swap the current module to the "real" Coveo variable.
-import _ = require('underscore');
-let swapVar = () => {
-if (window['Coveo'] == undefined) {
-window['Coveo'] = this;
-} else {
-_.each(_.keys(this), (k) => {
-window['Coveo'][k] = this[k];
-})
-}
-}
-swapVar();
+import {swapVar} from './SwapVar.ts';
+swapVar(this);
