+++ src/SearchboxIndex.ts
@@ -8,10 +8,13 @@ export * from './ModelsModules';
 export * from './UIBaseModules';
 export * from './TemplatesModules';
 
-export { Analytics } from './ui/Analytics/Analytics.ts';
-export { AnalyticsSuggestions } from './ui/AnalyticsSuggestions/AnalyticsSuggestions.ts';
-export { FieldSuggestions } from './ui/FieldSuggestions/FieldSuggestions.ts';
-export { Omnibox } from './ui/Omnibox/Omnibox.ts';
-export { Querybox } from './ui/Querybox/Querybox.ts';
-export { SearchButton } from './ui/SearchButton/SearchButton.ts';
-export { Searchbox } from './ui/Searchbox/Searchbox.ts';
+export {Analytics} from './ui/Analytics/Analytics';
+export {AnalyticsSuggestions} from './ui/AnalyticsSuggestions/AnalyticsSuggestions';
+export {FieldSuggestions} from './ui/FieldSuggestions/FieldSuggestions';
+export {Omnibox} from './ui/Omnibox/Omnibox';
+export {Querybox} from './ui/Querybox/Querybox';
+export {SearchButton} from './ui/SearchButton/SearchButton';
+export {Searchbox} from './ui/Searchbox/Searchbox';
+
+import {swapVar} from './SwapVar';
+swapVar(this);
