+++ test/ui/RecommendationTest.ts
@@ -164,14 +164,14 @@ export function RecommendationTest() {
         });
       });
 
-      it('exposes option autoTriggerQuery that should be set to false if there is a main search interface', ()=> {
+      it('exposes option autoTriggerQuery that should be set to false if there is a main search interface', () => {
         expect(options.mainSearchInterface).toBeDefined();
         options.autoTriggerQuery = true;
         test = Mock.optionsSearchInterfaceSetup<Recommendation, IRecommendationOptions>(Recommendation, options);
         expect(test.cmp.options.autoTriggerQuery).toBe(false);
       });
 
-      it('exposes options autoTriggerQuery that should be left as it is if there is no main search interface', ()=> {
+      it('exposes options autoTriggerQuery that should be left as it is if there is no main search interface', () => {
         options.mainSearchInterface = null;
         options.autoTriggerQuery = true;
         test = Mock.optionsSearchInterfaceSetup<Recommendation, IRecommendationOptions>(Recommendation, options);
